move some easy stuff
